[HOTSPOT-88] refactor: 공통 chart 컴포넌트 폴더 분리
