+++ packages/ui/src/components/index.ts
@@ -1,8 +1,8 @@
 export { Button } from './button/button';
 export { Card } from './card/card';
-export { LineChart } from './chart/LineChart';
-export { ProgressBar } from './chart/ProgressBar';
-export { DonutChartContainer } from './chart/DonutChartContainer';
+export { DonutChartContainer } from './chart/donut/DonutChartContainer';
+export { LineChart } from './chart/line/LineChart';
+export { ProgressBar } from './chart/progressbar/ProgressBar';
 export { Code } from './code/code';
 export { Slider } from './slider/Slider';
 export { Toggle } from './toggle/Toggle';
